+++ glue/viewers/scatter/layer_artist.py
@@ -12,11 +12,14 @@ from glue.core.exceptions import IncompatibleAttribute
 CMAP_PROPERTIES = set(['cmap_mode', 'cmap_att', 'cmap_vmin', 'cmap_vmax', 'cmap'])
 SIZE_PROPERTIES = set(['size_mode', 'size_att', 'size_vmin', 'size_vmax', 'size_scaling', 'size'])
 LINE_PROPERTIES = set(['linewidth', 'linestyle'])
+
 VISUAL_PROPERTIES = (CMAP_PROPERTIES | SIZE_PROPERTIES |
                      LINE_PROPERTIES | set(['color', 'alpha', 'zorder', 'visible']))
 
 DATA_PROPERTIES = set(['layer', 'x_att', 'y_att', 'cmap_mode', 'size_mode',
-                       'xerr_att', 'yerr_att', 'xerr_visible', 'yerr_visible'])
+                       'xerr_att', 'yerr_att', 'xerr_visible', 'yerr_visible',
+                       'vector_visible', 'vx_att', 'vy_att', 'vector_show_arrow',
+                       'origin_pos'])
 
 
 class InvertedNormalize(Normalize):
@@ -46,13 +49,18 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
         self.scatter_artist = self.axes.scatter([], [])
         self.plot_artist = self.axes.plot([], [], 'o', mec='none')[0]
         self.errorbar_artist = self.axes.errorbar([], [], fmt='none')
+        self.vector_artist = self.axes.quiver([], [], [], [], units='width',
+                                              pivot='mid',
+                                              headwidth=1, headlength=0) # x, y, vx, vy
 
         # Line
         self.line_artist = self.axes.plot([], [], '-')[0]
 
         self.mpl_artists = [self.scatter_artist, self.plot_artist,
-                            self.errorbar_artist, self.line_artist]
+                            self.errorbar_artist, self.line_artist,
+                            self.vector_artist]
         self.errorbar_index = 2
+        self.vector_index = 3
 
         self.reset_cache()
 
@@ -117,6 +125,14 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
                     except AttributeError:  # Matplotlib < 1.5
                         pass
 
+            if self.vector_artist is not None:
+                try:
+                    self.vector_artist.remove()
+                except ValueError:
+                    pass
+                except AttributeError:   # Matplotlib < 1.5
+                    pass
+
             if self.state.xerr_visible or self.state.yerr_visible:
 
                 if self.state.xerr_visible and self.state.xerr_att is not None:
@@ -132,6 +148,33 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
                 self.errorbar_artist = self.axes.errorbar(x, y, fmt='none',
                                                           xerr=xerr, yerr=yerr)
                 self.mpl_artists[self.errorbar_index] = self.errorbar_artist
+
+            if self.state.vector_visible:
+
+                if self.state.vx_att is not None and self.state.vy_att is not None:
+
+                    vx = self.layer[self.state.vx_att].ravel()
+                    vy = self.layer[self.state.vy_att].ravel()
+                    if self.state.vector_mode == 'Polarization':
+                        ang = vx
+                        length = vy
+                        # assume ang is anti clockwise from the x axis
+                        vx = length * np.cos(np.radians(ang))
+                        vy = length * np.sin(np.radians(ang))
+                else:
+                    vx = None
+                    vy = None
+
+                if self.state.vector_show_arrow:
+                    hw = 3
+                    hl = 5
+                else:
+                    hw = 1
+                    hl = 0
+                self.vector_artist = self.axes.quiver(x, y, vx, vy, units='width',
+                                                      pivot=str(self.state.origin_pos),
+                                                      headwidth=hw, headlength=hl)
+                self.mpl_artists[self.vector_index] = self.vector_artist
 
         elif self.state.style == 'Line':
 
@@ -180,17 +223,25 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
 
                     if self.state.cmap_mode == 'Fixed':
                         self.scatter_artist.set_facecolors(c)
+                        self.vector_artist.set_facecolors(c)
                     else:
                         self.scatter_artist.set_array(c)
                         self.scatter_artist.set_cmap(cmap)
+                        self.vector_artist.set_array(c)
+                        self.vector_artist.set_cmap(cmap)
                         if vmin > vmax:
                             self.scatter_artist.set_clim(vmax, vmin)
                             self.scatter_artist.set_norm(InvertedNormalize(vmax, vmin))
+                            self.vector_artist.set_clim(vmax, vmin)
+                            self.vector_artist.set_norm(InvertedNormalize(vmax, vmin))
                         else:
                             self.scatter_artist.set_clim(vmin, vmax)
                             self.scatter_artist.set_norm(Normalize(vmin, vmax))
+                            self.vector_artist.set_clim(vmin, vmax)
+                            self.vector_artist.set_norm(Normalize(vmin, vmax))
 
                     self.scatter_artist.set_edgecolor('none')
+                    self.vector_artist.set_edgecolor('none')
 
                 if force or any(prop in changed for prop in SIZE_PROPERTIES):
 
@@ -227,6 +278,20 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
 
                     if force or 'zorder' in changed:
                         eartist.set_zorder(self.state.zorder)
+
+            if self.state.vector_visible and self.vector_artist is not None:
+
+                if force or 'color' in changed:
+                    self.vector_artist.set_color(self.state.color)
+
+                if force or 'alpha' in changed:
+                    self.vector_artist.set_alpha(self.state.alpha)
+
+                if force or 'visible' in changed:
+                    self.vector_artist.set_visible(self.state.visible)
+
+                if force or 'zorder' in changed:
+                    self.vector_artist.set_zorder(self.state.zorder)
 
         elif self.state.style == 'Line':