+++ glue/viewers/scatter/qt/layer_style_editor.py
@@ -30,13 +30,17 @@ class ScatterLayerStyleEditor(QtWidgets.QWidget):
 
         self.layer_state.add_callback('xerr_visible', self._update_xerr_att_combo)
         self.layer_state.add_callback('yerr_visible', self._update_yerr_att_combo)
+        self.layer_state.add_callback('vector_visible', self._update_vector_att_combo)
         self.layer_state.add_callback('size_mode', self._update_size_mode)
+        self.layer_state.add_callback('vector_mode', self._update_vector_mode)
         self.layer_state.add_callback('cmap_mode', self._update_cmap_mode)
         self.layer_state.add_callback('layer', self._update_warnings)
 
         self._update_xerr_att_combo()
         self._update_yerr_att_combo()
+        self._update_vector_att_combo()
         self._update_size_mode()
+        self._update_vector_mode()
         self._update_cmap_mode()
         self._update_warnings()
 
@@ -45,6 +49,11 @@ class ScatterLayerStyleEditor(QtWidgets.QWidget):
 
     def _update_yerr_att_combo(self, *args):
         self.ui.combosel_yerr_att.setEnabled(self.layer_state.yerr_visible)
+
+    def _update_vector_att_combo(self, *args):
+        self.ui.combosel_vx_att.setEnabled(self.layer_state.vector_visible)
+        self.ui.combosel_vy_att.setEnabled(self.layer_state.vector_visible)
+        self.ui.bool_vector_show_arrow.setEnabled(self.layer_state.vector_visible)
 
     def _update_warnings(self):
 
@@ -68,6 +77,11 @@ class ScatterLayerStyleEditor(QtWidgets.QWidget):
         else:
             self.ui.label_warning_errorbar.hide()
 
+        if n_points > 10000:
+            self.ui.label_warning_vector.show()
+        else:
+            self.ui.label_warning_vector.hide()
+
     def _update_size_mode(self, size_mode=None):
 
         if self.layer_state.size_mode == 'Fixed':
@@ -78,6 +92,15 @@ class ScatterLayerStyleEditor(QtWidgets.QWidget):
             self.ui.value_size.hide()
             self.ui.combosel_size_att.show()
             self.ui.size_row_2.show()
+
+    def _update_vector_mode(self, vector_mode=None):
+        if self.layer_state.vector_mode == 'Cartesian':
+            self.ui.label_vector_x.setText('vx')
+            self.ui.label_vector_y.setText('vy')
+        elif self.layer_state.vector_mode == 'Polarization':
+            self.ui.label_vector_x.setText('ang(deg)')
+            self.ui.label_vector_y.setText('length')
+
 
     def _update_cmap_mode(self, cmap_mode=None):